requirements.txt
fastapi
uvicorn[standard]
pydantic-settings
python-dotenv
sqlalchemy
pymysql
cryptography
psycopg2-binary
faker
